requirements.txt
# Agency-in-a-Box Python dependencies — ALL pinned (spec requirement)

# Database
psycopg2-binary==2.9.9

# AI Models
anthropic==0.39.0
google-genai>=1.66.0
instructor>=1.14.0

# Retry logic
tenacity==9.0.0

# Pydantic
pydantic==2.10.3

# HTTP client
httpx==0.28.1

# Telegram
python-telegram-bot==21.7

# DMZ FastAPI
fastapi==0.115.6
uvicorn==0.34.0
slowapi==0.1.9

# Linting/QA (for local dev — also pinned in Dockerfile)
ruff==0.4.9
mypy==1.10.0
bandit==1.7.9
pytest==8.2.2
pytest-cov==5.0.0
pytest-json-report==1.5.0
